insightai-api/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pydantic==2.7.4
python-multipart==0.0.9
httpx==0.27.0
pymupdf==1.24.5
pdfminer.six==20231228
python-dotenv==1.0.1
